moving data-export-functions to own service #2405
