+++ test/unit/mock/services/data-export.js
@@ -1,0 +1,13 @@
+module.exports = [function () {
+    return {
+        prepareExport: function () {},
+        requestExport: function () {
+            return true;
+        },
+        showCSVResults: function () {
+            return true;
+        },
+        handleArrayBuffer: function () {},
+        loadingStatus: function () {}
+    };
+}];
